Expose CardSearchDelegate
